requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.0.6
cycler==0.11.0
Django==5.0.6
django-ckeditor==6.7.1
django-debug-toolbar==4.3.0
django-environ==0.8.0
django-filter==24.2
django-js-asset==2.2.0
django-taggit==5.0.1
djangorestframework==3.15.1
easy-thumbnails==2.8.5
fonttools==4.38.0
frozenlist==1.4.1
geoip2==4.8.0
idna==3.7
kiwisolver==1.4.4
Markdown==3.6
matplotlib==3.6.2
maxminddb==2.6.2
multidict==6.0.5
numpy==1.23.5
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.32.3
scipy==1.9.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
